backend/pyproject.toml
[project]
name = "Hexagon-AI"
version = "0.1.0"
description = "Ai Interviewer for Hexagon-A"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp[speedups]>=3.9.3",
    "fastapi[standard]>=0.115.8",
    "google-generativeai>=0.8.4",
    "openai>=1.59",
    "pipecat-ai[silero,daily,elevenlabs,google,deepgram,openai,websocket,cartesia]>=0.0.67",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
    "loguru>=0.7.3",
    "supabase>=1.0.3",
    "pipecat-ai-flows>=0.0.15",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pre-commit>=3.5.0"
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "C4", "ARG", "SIM"]
ignore = []
fixable = ["ALL"]

[tool.ruff.lint.isort]
force-single-line = false
force-wrap-aliases = true
force-sort-within-sections = true
combine-as-imports = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
skip-magic-trailing-comma = false
